[FLINK-25432][runtime] Moves cleanup test from DispatcherTest into DispatcherResourceCleanupTest
